use parameter in get api response
