+++ src/useAsync.js
@@ -26,6 +26,7 @@ function reducer(state, { type, data, error }) {
 }
 
 function useAsync(callback, deps=[], skip=false) {
+  
   const [state, dispatch] = useReducer(reducer, {
     loading: false,
     data: null,
